logiflow/mvp2-analytics-layer/2A-dashboard/requirements.txt
streamlit==1.32.0
pandas
sqlalchemy
psycopg2-binary
python-dotenv
plotly==5.20.0